+++ examples/modules/super_chain_demo_b/models/__init__.py
@@ -1,0 +1,1 @@
+from . import order  # noqa: F401